bra va

+++ src/server/common/TimeMapper.h
@@ -31,8 +31,8 @@ public:
     return 0 <= index && index < sampleCount? index : -1;
   }
 
-  TimeStamp unmap(int i) const {
-    return offset + double(i)*period;
+  TimeStamp unmap(double i) const {
+    return offset + i*period;
   }
 };
